mfdb_sql: Don't try checking if existing function is aggregate
Postgres 11 replaces proisagg with prokind without backwards compatibility.
However, as we manage the schema there should be no reason for a matching
function not to be an aggregate function, and even if it was the following code
would error anyway, as the function would be in the way. Remove check