mfdb_aggregate_group: Include col in temporary table name
Allow pre_query to modify groups as they go through, use this to
include the column in the temporary table name.

In theory this is a decent plan, but it makes the machinations to
preserve temporary tables pointless, since the user copy of the
variable won't have the name in.