ewe_model: Rewrite Rpath support to use Rpath::create.rpath.params
Rpath now has more useful model data handling built-in, although the
general principles are still the same. Use the methods it provides
instead of handing out raw data.

Since Rpath isn't on CRAN yet, we can't depend directly on it. To bodge
around this but keep check --as-cran passing, the
Rpath::create.rpath.params function needs to be passed in for now.