mfdb_aggregate_group: Don't require divisions when grouping by areacell
Before a division had to be made, even when they weren't being used.
The general advice was to make a division with a matching name to the
areacell it contained.

This is a bit rubbish, and confusing in comparison to every other
taxonomy, where you can use a name or a t_group.

Query the combination of areacell and division, so divisions aren't
required unless you need them.